LPS-79878 This verification was added by LPS-49661 in 7.0. Moved to an upgrade step for that the upgrade to 7.0
